requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
djangorestframework
gunicorn
idna
packaging
pytz
requests
sqlparse
tzdata
urllib3